remove some more deprecation warnings
