+++ DQMOffline/CalibTracker/plugins/SealModules.cc
@@ -1,7 +1,4 @@
 #include "FWCore/Framework/interface/MakerMacros.h"
 
-#include "DQMOffline/CalibTracker/plugins/SiStripBadComponentsDQMServiceReader.h"
-DEFINE_FWK_MODULE(SiStripBadComponentsDQMServiceReader);
-
 #include "DQMOffline/CalibTracker/plugins/SiStripQualityHotStripIdentifierRoot.h"
 DEFINE_FWK_MODULE(SiStripQualityHotStripIdentifierRoot);
