+++ EventFilter/SiStripRawToDigi/test/plugins/SiStripModuleTimer.cc
@@ -1,6 +1,6 @@
 // user includes
 #include "FWCore/ServiceRegistry/interface/Service.h"
-#include "FWCore/Framework/interface/EDAnalyzer.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/ParameterSet/interface/ParameterSet.h"
 //#include "DQM/HLTEvF/interface/PathTimerService.h"
@@ -9,7 +9,7 @@
 #include "TFile.h"
 #include "TTree.h"
 
-class SiStripModuleTimer : public edm::EDAnalyzer {
+class SiStripModuleTimer : public edm::one::EDAnalyzer<> {
 public:
   SiStripModuleTimer(const edm::ParameterSet&);
   ~SiStripModuleTimer();
